feat: add prepare_install methods
Those methods are used to get the source code, archives or binaries of the different tools to be installed. All downloaded sources are stored in SIU/archives.
This allows to perform an offline installation : download all using a prepare_install, copy the SIU folder then perform the install.
An online installation performs a prepare_install then the install.